Revert "Introduce common test base for parser tests"
This reverts commit 3cac506c8c0e3fdd082a9906385d2eb6e6c087c6.
