+++ spring-data-jpa/src/test/java/org/springframework/data/jpa/repository/query/JpqlComplianceTests.java
@@ -15,31 +15,51 @@
  */
 package org.springframework.data.jpa.repository.query;
 
-import java.util.List;
+import static org.assertj.core.api.Assertions.*;
+import static org.springframework.data.jpa.repository.query.JpaQueryParsingToken.*;
 
 import org.antlr.v4.runtime.CharStreams;
 import org.antlr.v4.runtime.CommonTokenStream;
-import org.antlr.v4.runtime.tree.ParseTree;
+import org.junit.jupiter.api.Test;
 
 /**
  * @author Christoph Strobl
  */
-public class JpqlComplianceTests extends SqlParserTests {
+public class JpqlComplianceTests {
 
-	@Override
-	ParseTree parse(String query) {
+	private static String parseWithoutChanges(String query) {
 
 		JpqlLexer lexer = new JpqlLexer(CharStreams.fromString(query));
 		JpqlParser parser = new JpqlParser(new CommonTokenStream(lexer));
 
 		parser.addErrorListener(new BadJpqlGrammarErrorListener(query));
 
-		return parser.start();
+		JpqlParser.StartContext parsedQuery = parser.start();
+
+		return render(new JpqlQueryRenderer().visit(parsedQuery));
 	}
 
-	@Override
-	<T extends ParseTree> List<JpaQueryParsingToken> analyze(T parseTree) {
-		return new JpqlQueryRenderer().visit(parseTree);
+	private void assertQuery(String query) {
+
+		String slimmedDownQuery = reduceWhitespace(query);
+		assertThat(parseWithoutChanges(slimmedDownQuery)).isEqualTo(slimmedDownQuery);
+	}
+
+	private String reduceWhitespace(String original) {
+
+		return original //
+				.replaceAll("[ \\t\\n]{1,}", " ") //
+				.trim();
+	}
+
+	@Test
+	void numericLiterals() {
+
+		assertQuery("SELECT e FROM  Employee e WHERE e.id = 1234");
+		assertQuery("SELECT e FROM  Employee e WHERE e.id = 1234L");
+		assertQuery("SELECT s FROM  Stat s WHERE s.ratio > 3.14");
+		assertQuery("SELECT s FROM  Stat s WHERE s.ratio > 3.14F");
+		assertQuery("SELECT s FROM  Stat s WHERE s.ratio > 3.14e32D");
 	}
 
 }
